+++ aws/session/testdata/credential_source_config
@@ -1,16 +1,20 @@
 [env_var_credential_source]
-role_arn = arn
+role_arn = assume_role_w_creds_role_arn_env
 credential_source = Environment
 
 [invalid_source_and_credential_source]
-role_arn = arn
+role_arn = assume_role_w_creds_role_arn_bad
 credential_source = Environment
 source_profile = env_var_credential_source
 
 [ec2metadata]
-role_arn = assume_role_w_creds_role_arn
+role_arn = assume_role_w_creds_role_arn_ec2
 credential_source = Ec2InstanceMetadata
 
 [ecscontainer]
-role_arn = assume_role_w_creds_role_arn
+role_arn = assume_role_w_creds_role_arn_ecs
 credential_source = EcsContainer
+
+[chained_assume_role]
+role_arn = assume_role_w_creds_role_arn_chain
+source_profile = ec2metadata
